+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
   * Avoid modifying frozen strings ([yuri-zubov][yuri-zubov])
+  * Define `#respond_to_missing?` on `Capybara::Node::Email`
+    ([tylerhunt][tylerhunt])
 
 ## 3.0.0
 
@@ -17,4 +19,5 @@
   * Corrects `inspect` of `Capybara::Node::Email`
   * Delegate all missing methods in `Capybara::Node::Email` to `base.email`
 
+[tylerhunt]: https://github.com/tylerhunt
 [yuri-zubov]: https://github.com/yuri-zubov